Extract Diagnostic::Helper
